refactored sentiment-prompt
